backend/requirements.txt
fastapi[all]
uvicorn
python-multipart
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai
PyMuPDF
psycopg2-binary
pydantic>=2.0
python-dotenv